feat: add ts-loader to Webpack and migrate first class to TypeScript. Changing extension from .js to .ts seems to break history/diffs
